requirements.txt
aioflask==0.4.0
APScheduler==3.9.1
df2img==0.2.2
Flask-Cors==1.10.3
httpx==0.22.0
ipython==7.30.1
nbformat==5.1.3
nextcord==2.0.0a8
pandas==1.4.1
Pyrebase4==4.5.0
yt-dlp==2022.2.4